One final tweak
Today's game was pretty weak. I still have a lot of homework to do
tonight so I'm going to cut this one off here.

What I learned: Cartoony VR can actually be more believable than
"realistic" vr.  Also, even though I usually frown on the
GameObject.Find method because it is so slow, today it saved me a lot of
dragging and dropping in the inspector. From now on I won't be so quick
to dismiss the GameObject.Find method for prototyping.